Add files via upload
The objective of this project is to analyze a year-long pizza sales dataset to derive actionable insights into customer behavior, sales trends, product performance, and time-based ordering patterns. The end goal is to build a fully interactive dashboard using Excel 2019 that is powered by SQL Server queries and pivot-based visualizations.